Fix bug in HTrigger configuration
The horizonal trigger configuration message was offset by a byte, and
this caused incorrect triggering behaviour. We now get a trigger message
fromt he scope.

+++ hantek_priv.h
@@ -36,7 +36,7 @@ struct hantek_channel {
     /**
      * The "level" of this channel
      */
-    uint8_t level;
+    uint16_t level;
 
     /**
      * Volts per division for this channel
